feat(swap): implement simulated disk swap space and complete virtual memory system (Phase 8)
- Add swap_disk_t module for simulated disk storage and slot bitmap management
- Extend page table entry with PAGE_FLAG_SWAPPED and swap_page_id tracking
- Integrate Swap Out and Swap In mechanisms into handle_page_fault handler
- Update page replacement and TLB invalidation workflows on page eviction
- Add swap unit tests (SwapTest) and update existing tests for 7-argument page fault handler

+++ tests/test_mmu.c
@@ -51,7 +51,7 @@ int main(void)
         "TLB Statistics | Hits: %llu | Misses: %llu | Hit Rate: %.2f%%",
         (unsigned long long)tlb.hits,
         (unsigned long long)tlb.misses,
-        tlb_get_hit_rate(&tlb));
+        (double)tlb_get_hit_rate(&tlb));
 
     tlb_destroy(&tlb);
     page_table_destroy(&pt);
